+++ americanoforum/src/forum/server/domainlayer/PersistenceDataHandler.java
@@ -15,9 +15,9 @@ public interface PersistenceDataHandler {
 
     Forum getForumFromXml();
     void addRegUserToXml(String username, String password, String email, String firstname, String lastname, String address, String gender, String up);
-    void addMsgToXml(String sbj, String cont, int msg_id, int parent_id, String username, Date datetime);
-    void modifyMsgInXml(int id_toChange, String newCont);
+    void addMsgToXml(String sbj, String cont, long msg_id, long parent_id, String username, Date datetime);
+    void modifyMsgInXml(long id_toChange, String newCont);
     void changeUserPermission(String username, String permission);
-    void deleteMsgFromXml(int msg_id);
+    void deleteMsgFromXml(long msg_id);
 
 }
