+++ americanoforum/src/forum/tcpcommunicationlayer/DeleteMessageMessage.java
@@ -12,10 +12,10 @@ import forum.server.domainlayer.ForumFacade;
  * @author Ilya
  */
 public class DeleteMessageMessage implements ClientMessage{
-   int _idToDelete;
+   long _idToDelete;
    String _user;
 
-    public DeleteMessageMessage(int _idToDelete, String _user) {
+    public DeleteMessageMessage(long _idToDelete, String _user) {
         this._idToDelete = _idToDelete;
         this._user = _user;
     }
